requirements.txt
redis
requests
jira
pytest
google-cloud-secret-manager
google-cloud-aiplatform
google-cloud-storage
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-exporter-gcp-trace
